Add BaseLevelDB class
